Created new scanner
The previous lexical analysis was too slow ( it used REGEX ) to be able to run in between keystrokes. Using Instruments to analyse it showed that it was the REGEX itself that was too slow and was causing the whole scanning process to be too slow. I removed the REGEX and rewrote the scanner to use an improved version of the scanner I had written for ArgonWorks. Just tested the scanner now and it's super fast and easily manages to scan the whole source file between keystrokes. Works really well. Still has the odd glitch but it should be easy to fix any bugs.