Improved scanning +  UI
Improved scanning to make locations more accurate. Fixed bugs in BracketMatcher which now works correctly. Also fixed bugs in scanner to do with handing the end of source correctly. Changed sizing of the PathControl and moved the issues field into the PathControl since it looks better. Still having problems sizing the split view correctly when the window is opened.