Reworked Project Editor
Changed editing so that it handles both Visual Design editing and Source Code editing. Still have bugs in ProjectSourceViewController that need to be fixed. InspectorViewController still unimplemented. Current problem is that the ruler view is drawing line numbers to infinity, need to find a way to fix that. Once editing fixed will proceed with compilation.